Launches: load past and upcoming launches

+++ app/src/main/java/com/github/emilg1101/spacex/data/spacexapi/pojo/response/Payload.kt
@@ -10,7 +10,7 @@ data class Payload(
     @SerializedName("nationality") val nationality: String,
     @SerializedName("manufacturer") val manufacturer: String,
     @SerializedName("payload_type") val payloadType: String,
-    @SerializedName("payload_mass_kg") val payloadMassKg: Int,
+    @SerializedName("payload_mass_kg") val payloadMassKg: Double,
     @SerializedName("payload_mass_lbs") val payloadMassLbs: Double,
     @SerializedName("orbit") val orbit: String,
     @SerializedName("orbit_params") val orbitParams: OrbitParams
